+++ auth/index.ts
@@ -1,2 +1,2 @@
-export { UserAccess, accessLevels } from './roles';
+export { UserAccess, accessLevels } from './access';
 export { User } from './user';
